xagent/xagent/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
cryptography==41.0.7
psutil==5.9.6
redis==5.0.1
sentry-sdk[fastapi]==1.38.0
requests==2.31.0
httpx==0.25.2
pinecone-client==3.0.0
openai==1.3.0
playwright==1.40.0